+++ apps/dsp-app/src/app/system/projects/projects.component.ts
@@ -10,12 +10,12 @@ import {
     UserResponse,
 } from '@dasch-swiss/dsp-js';
 import { DspApiConnectionToken } from '@dasch-swiss/vre/shared/app-config';
-import { DialogComponent } from '@dsp-app/src/app/main/dialog/dialog.component';
 import { AppErrorHandler } from '@dasch-swiss/vre/shared/app-error-handler';
 import {
     Session,
     SessionService,
 } from '@dasch-swiss/vre/shared/app-session';
+import { AppProgressIndicatorComponent } from '@dasch-swiss/vre/shared/app-progress-indicator';
 
 /**
  * projects component handles the list of projects
@@ -132,23 +132,6 @@ export class ProjectsComponent implements OnInit {
                     }
                 );
         }
-    }
-
-    openDialog(mode: string): void {
-        const dialogConfig: MatDialogConfig = {
-            width: '560px',
-            maxHeight: '80vh',
-            position: {
-                top: '112px',
-            },
-            data: { mode: mode },
-        };
-
-        const dialogRef = this._dialog.open(DialogComponent, dialogConfig);
-
-        dialogRef.afterClosed().subscribe(() => {
-            // update the view
-        });
     }
 
     /**