+++ apps/dsp-app/src/app/user/overview/overview.component.ts
@@ -193,4 +193,8 @@ export class OverviewComponent implements OnInit {
                 break;
         }
     }
+
+    createNewProject() {
+        this._router.navigate(['project', 'create-new']);
+    }
 }