+++ apps/dsp-app/src/app/workspace/resource/directives/text-value-html-link.directive.ts
@@ -9,26 +9,62 @@ export class TextValueHtmlLinkDirective {
     @Output() internalLinkHovered = new EventEmitter<string>();
 
     /**
-     * react to a click event for an internal link.
+     * react to a click event for a link (left mouse button only). If it is an internal link, emit the event.
+     * If it is an external link, open it in a new tab.
      *
-     * @param targetElement the element that was clicked.
+     * @param event: The mouse event.
      */
-    @HostListener('mousedown', ['$event.target'])
-    onClick(targetElement) {
-        if (targetElement.nodeName.toLowerCase() !== 'a') {
-            return false; // only handle click events on links
-        }
+    @HostListener('click', ['$event'])
+    onClick(event: MouseEvent) {
+        const targetElement = event.target as HTMLAnchorElement;
         if (
-            targetElement.className
+            !targetElement ||
+            targetElement?.nodeName?.toLowerCase() !== 'a' ||
+            event.button === 2 || // right mouse button
+            event.button === 1 // middle mouse button
+        ) {
+            // only handle left mouse click events on links, all other events are ignored here
+            return;
+        }
+        if ( targetElement?.className
                 .toLowerCase()
                 .indexOf(Constants.SalsahLink) !== -1
-        ) {
+        ) { // if it is an internal link, override all behaviour, so also second mouse button clicks, hence the iris
+            // like "rdfh.ch/ ..." can not be dereferenced outside the app by a browser
             this.internalLinkClicked.emit(targetElement.href);
-            return false;
-        } else {
-            // open all other links as external links in a new tab
+            // prevent the default action for internal links
+            event.preventDefault();
+            return;
+
+        } else { // left mouse clicks on external links
+            // open in a new tab
             window.open(targetElement.href, '_blank');
-            return false;
+            event.preventDefault();
+        }
+    }
+
+    /**
+     * react to a mouse down event for a link (middle and second mouse button).
+     * For middle mouse and second mouse button clicks there is need to
+     * handle mousedown events instead of click events in order to override browsers default behaviour
+     * (open in new tab, context menu). The default behaviour is prevented for internal links.
+     *
+     * @param event: The mouse event.
+     */
+    @HostListener('mousedown', ['$event'])
+    onMouseDown(event: MouseEvent) {
+        const targetElement = event.target as HTMLAnchorElement;
+
+        if (!targetElement || targetElement.nodeName.toLowerCase() !== 'a') {
+            return; // only handle mouse events on links
+        }
+
+        // handle middle mouse and second mouse clicks here
+        if (event.button === 1 || event.button === 2) {
+            if (targetElement.className.toLowerCase().indexOf(Constants.SalsahLink) !== -1) {
+                this.internalLinkClicked.emit(targetElement.href);
+                event.preventDefault();
+            }
         }
     }